Completed refactor and deleted duplicate files.
