--- client_new
@@ -1,62 +1,0 @@
-import socket
-import os
-import json
-
-
-IP = socket.gethostbyname(socket.gethostname())  # get the server IP address
-PORT = 2233                                      # port number
-ADDR = (IP, PORT)                                # for the address IP and PORT
-SIZE = 1024                                      # for the size
-FORMAT = "utf-8"                                 # for encode data
-
-# Dictionary data
-a_dict = {'Common Name': 'Otter', 'Scientific Name': 'Mustelidae', 'Type': 'Mammal', 'Size': '2 to 6 feet',
-          'Weight': '10 to 75 pounds'}
-
-# To convert python dictionary to JSON
-json_object = json.dumps(a_dict, indent=4)
-
-# Create the data directory if not exists
-try:
-    os.mkdir("data_files/", mode=0o777)
-except OSError as error:
-    print()
-
-"""Create the file and write the JSON object """
-file_directory = "data_files/"
-file_name = "dict3.txt"
-File_Path_Name = os.path.join(file_directory, file_name)
-keyfile_name = os.path.join(file_directory, "keyfile.ky")
-f = open(File_Path_Name, 'w')  # open file in write mode
-f.write(json_object)
-f.close()
-
-
-def main():
-    """ Open and read the file """
-
-
-file = open(File_Path_Name, "r")
-data = file.read()
-
-print(f"[CLIENT] Data : " + data)
-""" Staring TCP socket """
-client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-""" Connecting to the server """
-client.connect(ADDR)
-print(ADDR)
-""" Sending the filename to the server. """
-client.send(file_name.encode(FORMAT))
-print(f"[CLIENT] File Sent to the server ...")
-msg = client.recv(SIZE).decode(FORMAT)
-print(f"[SERVER]: {msg}")
-""" Sending the file data to the server. """
-client.send(data.encode(FORMAT))
-msg = client.recv(SIZE).decode(FORMAT)
-print(f"[SERVER]: {msg}")
-""" Closing the file. """
-file.close()
-""" Closing the connection from the server. """
-client.close()
-if __name__ == "__main__":
-    main()
