--- server_new
@@ -1,49 +1,0 @@
-import socket
-
-IP = socket.gethostbyname(socket.gethostname())  # get the server IP address
-PORT = 2233                                      # port number
-ADDR = (IP, PORT)                                # for the address IP and PORT
-SIZE = 1024                                      # for the size
-UTF_FORMAT = "utf-8"                             # for decode data
-
-
-def main():
-    """ Staring Server """
-
-
-print("[STARTING] Server is starting." + IP)
-""" Staring TCP socket """
-server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-""" Bind the IP and PORT  """
-server.bind(ADDR)
-""" Server is listening, it's now waiting for the client to connected """
-server.listen(5)
-print("[LISTENING] Server is listening.")
-""" To accept the connection from the client """
-conn, addr = server.accept()
-print(f"[NEW CONNECTION] {addr} connected.")
-print(f"[CLIENT] Receiving the file from the client.")
-""" Receiving the filename from the client """
-filename = conn.recv(SIZE).decode(UTF_FORMAT)
-""" Send filename received to the client """
-conn.send("Filename received.".encode(UTF_FORMAT))
-print(f"[Client] Filename is : " + filename)
-""" Open a copy of the filename """
-file = open("copy_" + filename, "w")
-print(f"[Client] Receiving the data from the client.")
-""" Receiving the data from the client """
-data = conn.recv(SIZE).decode(UTF_FORMAT)
-print(f"[SERVER] Data " + data + " is Received ...")
-""" Send the data is received from the client """
-conn.send("Data received from the server.".encode(UTF_FORMAT))
-""" Write the data to the copy file """
-file.write(data)
-print("[SERVER] Successfully Completed.")
-
-""" Closing the file. """
-file.close()
-""" Closing the connection from the client. """
-conn.close()
-print(f"[DISCONNECTED] {addr} disconnected.")
-if __name__ == "__main__":
-    main()
